requirements.txt
Babel==2.3.4
Django==1.10.1
django-phonenumber-field==1.1.0
phonenumberslite==7.7.1
pkg-resources==0.0.0
pytz==2016.6.1
